Add style change when buttons are click or push
This helps to recreate a button push animation, whether the button is
clicked in the ui or their related key is pushed on the keyboard. These
test are being implemented by sections, as all the logic behind these
functionalities prove to be effective. They are going to be added to
all buttons.

The result() function from the equal button, now accepts both keydown
& click events to be activated.