Fix bugs with +/- btn & begin refactoring fn names
There were 2 bugs identified. The first one led to the sign change of
the first number, when no other negative number was present in the
expression. The second one, inputted the minus sign when no number
was present.

I changed the order of some conditional statements and modified some
conditions being evaluated by them. Now the +/- btn work as expected:
changes the sign of only the current number being operated on & do not
work on an empty display.